@@ -1,10 +1,10 @@
 name: factory_castle
 description: IoC container inspired by Castle Windsor + Flutter MVVM framework. Your whole app architecture in a single package you've been dreaming about.
-version: 2.0.2
+version: 3.0.0
 repository: https://github.com/nuc134r/factory_castle
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
   test: ^1.15.5
